Soem more C2S stuff moved
